Update Tron.php
$receiver_address was added to the parameters of freezeBalance and unfreezeBalance.  The Tron developers documentation has a section where it explains how to freeze and unfreeze TRX. 

The section of FreezeBalance describes that you can freeze an amount of TRX. Will give bandwidth OR Energy and TRON Power (voting rights) to the owner of the frozen tokens. Optionally, can freeze TRX to grant Energy or Bandwidth to other users. Balance amount in the denomination of Sun.  The parameter for receiver_address is the address that will receive the resource.

The section of UnfreezeBalance describes TRX that has passed the minimum freeze duration. Unfreezing will remove bandwidth and TRON Power. Returns unfrozen TRX transaction. The parameter for receiver_address is the address that will lose the resource.